refactor(docs-refresh): split into types/markers/io modules
Brings packages/crane-mcp/src/cli/docs-refresh.ts under structural caps.

- docs-refresh.ts (~120 code lines): thin orchestrator. main() split into
  runAudit, runInitMarkers, runRefreshTarget, runRefresh sub-functions,
  each under 75 lines.
- docs-refresh-types.ts: types, interfaces, constants (MARKED_PAGES,
  ALL_PAGE_TYPES, INIT_ANCHORS, regex, body format constants).
- docs-refresh-markers.ts: parseMarkers, diffGate, replaceBlockBody,
  isBulletListBody, initMarkersForPage, initMarkersForPageDetailed +
  tryWrapAfterHeading, appendAsNewSection, stripBlocks helpers.
- docs-refresh-io.ts: ghFetcher, renderers, loadRefreshConfig,
  auditAllVenturePages, formatAuditHuman, expandScopes, parseArgs,
  refreshOnePage + writeRefreshedPage, warnMissingMarkers helpers.

main() complexity 22 → split per-mode. Public API preserved. 764/764 tests pass.
